Broke sections into separate components, added submitted notice
