+++ app/assets/stylesheets/end-of-cycle.scss
@@ -1,0 +1,73 @@
+.end-of-cycle-max {
+  text-align: center;
+  font-size: 60px;
+  font-weight: bold;
+  border: solid #3d95bf 2px;
+  margin: 0px 0px 0px 0px;
+}
+
+.end-of-cycle-max-right {
+  text-align: center;
+  font-size: 60px;
+  font-weight: bold;
+  border: solid #3d95bf 2px;
+  margin: 0px 0px 0px 20px;
+}
+
+.eoc-title {
+  font-size: 30px;
+  font-weight: bold;
+  color: #F55D3E;
+  margin-bottom: 0px;
+  text-align: center;
+}
+
+.eoc-text {
+  text-align: center;
+}
+
+.eoc-plot-display {
+  border: solid #1a2e38 4px;
+  padding: 20px 20px 20px 20px;
+  box-shadow:
+    inset 0 0 0 5px #1a2e38,
+    inset 0 0 0 12px white;
+  background-color: white;
+}
+
+.eoc-plot-display-2 {
+  border: solid #1a2e38 4px;
+  padding: 50px 20px 20px 20px;
+  box-shadow:
+    inset 0 0 0 5px #1a2e38,
+    inset 0 0 0 12px white;
+  background-color: white;
+}
+
+.eoc-left {
+  padding-right: 20px;
+}
+
+.eoc-congrats-title {
+  font-size: 40px;
+  font-weight: bold;
+  margin-top: 30px;
+  margin-bottom: 30px;
+  margin-left: 2%;
+  margin-right: 2%;
+  color: #F55D3E;
+}
+
+.eoc-small-text {
+  font-size: 14px;
+  line-height: 14px;
+  margin-bottom: 5px;
+  margin-top: 28px;
+}
+.eoc-small-text-2 {
+  font-size: 14px;
+  line-height: 14px;
+  margin-bottom: 5px;
+  margin-top: 28px;
+  margin-left: 20px;
+}
